Use the App plugin via the new backend system
